Commit the first version of the AES70 compliancy test tool
This (open source) version of the AES70 compliancy test tool (2.0.0) replaces the previous version of the compliancy test tool (Which was only available for member companies). 
This version boths checks for the AES70-2015 version, and it contains preliminary checks for the AES70-2018 version. 

Some house-keeping has to be done in the repo.